refactor(auction): AuctionEndedEvent 공유 폴더로 이동
